+++ compiler/lib-runtime-files/js_of_ocaml_compiler_runtime_files.ml
@@ -34,7 +34,6 @@ let runtime =
     ; hash
     ; ieee_754
     ; int64
-    ; internalMod
     ; ints
     ; io
     ; jslib