+++ compiler/lib/ocaml_compiler.ml
@@ -24,8 +24,7 @@ let rec constant_of_const c : Code.constant =
   match c with
   | Const_base (Const_int i) -> Int (Targetint.of_int_warning_on_overflow i)
   | Const_base (Const_char c) -> Int (Targetint.of_int_exn (Char.code c))
-  | ((Const_base (Const_string (s, _))) [@if ocaml_version < (4, 11, 0)])
-  | ((Const_base (Const_string (s, _, _))) [@if ocaml_version >= (4, 11, 0)]) -> String s
+  | Const_base (Const_string (s, _, _)) -> String s
   | Const_base (Const_float s) -> Float (float_of_string s)
   | Const_base (Const_int32 i) -> (
       match Config.target () with
@@ -40,8 +39,6 @@ let rec constant_of_const c : Code.constant =
   | Const_float_array sl ->
       let l = List.map ~f:(fun f -> float_of_string f) sl in
       Float_array (Array.of_list l)
-  | ((Const_pointer i) [@if ocaml_version < (4, 12, 0)]) ->
-      Int (Targetint.of_int_warning_on_overflow i)
   | Const_block (tag, l) ->
       let l = Array.of_list (List.map l ~f:constant_of_const) in
       Tuple (tag, l, Unknown)